+++ src/Lucene.Net.CodeAnalysis.Dev/LuceneDev1xxx/LuceneDev1000_FloatingPointEqualityCSCodeAnalyzer.cs
