+++ src/Lucene.Net.CodeAnalysis.Dev/LuceneDev1xxx/LuceneDev1001_FloatingPointFormattingCSCodeAnalyzer.cs
