+++ src/Lucene.Net.CodeAnalysis.Dev/LuceneDev1xxx/LuceneDev1002_FloatingPointArithmeticCSCodeAnalyzer.cs
