+++ src/Lucene.Net.CodeAnalysis.Dev/LuceneDev1xxx/LuceneDev1004_ArrayMethodReturnValueCSCodeAnalyzer.cs
