+++ tests/Lucene.Net.CodeAnalysis.Dev.CodeFixes.Tests/LuceneDev1xxx/TestLuceneDev1001_FloatingPointFormattingCSCodeFixProvider.cs
