+++ tests/Lucene.Net.CodeAnalysis.Dev.Tests/LuceneDev1xxx/TestLuceneDev1002_FloatingPointArithmeticCSCodeAnalyzer.cs
@@ -21,17 +21,18 @@ using Microsoft.CodeAnalysis.Testing;
 using NUnit.Framework;
 using System.Threading.Tasks;
 
-namespace Lucene.Net.CodeAnalysis.Dev.Tests
+namespace Lucene.Net.CodeAnalysis.Dev
 {
-    public class TestLuceneDev1001_FloatingPointFormattingCSCodeAnalyzer
+    public class TestLuceneDev1002_FloatingPointArithmeticCSCodeAnalyzer
     {
+
         //No diagnostics expected to show up
         [Test]
         public async Task TestEmptyFile()
         {
             var testCode = @"";
 
-            var test = new InjectableCSharpAnalyzerTest(() => new LuceneDev1001_FloatingPointFormattingCSCodeAnalyzer())
+            var test = new InjectableCSharpAnalyzerTest(() => new LuceneDev1002_FloatingPointArithmeticCSCodeAnalyzer())
             {
                 TestCode = testCode
             };
@@ -45,31 +46,38 @@ namespace Lucene.Net.CodeAnalysis.Dev.Tests
             var testCode = @"
         using System;
         using System.Collections.Generic;
-        using System.Globalization;
         using System.Linq;
         using System.Text;
+        using System.Threading.Tasks;
         using System.Diagnostics;
 
         public class MyClass
         {
             private readonly float float1 = 1f;
+            private readonly float float2 = 3.14f;
 
             public void MyMethod()
             {
-                string result = float1.ToString(CultureInfo.InvariantCulture);
+                long foo = 33;
+                var result = ((double)float1 * (double)float2) / foo;
             }
        }
        ";
 
-            var expected = DiagnosticResult.CompilerWarning(Descriptors.LuceneDev1001_FloatingPointFormatting.Id)
-                .WithMessageFormat(Descriptors.LuceneDev1001_FloatingPointFormatting.MessageFormat)
-                .WithArguments("float1.ToString")
-                .WithLocation("/0/Test0.cs", line: 15, column: 33);
+            var expected1 = DiagnosticResult.CompilerWarning(Descriptors.LuceneDev1002_FloatingPointArithmetic.Id)
+                .WithMessageFormat(Descriptors.LuceneDev1002_FloatingPointArithmetic.MessageFormat)
+                .WithArguments("((double)float1 * (double)float2) / foo")
+                .WithLocation("/0/Test0.cs", line: 17, column: 30);
 
-            var test = new InjectableCSharpAnalyzerTest(() => new LuceneDev1001_FloatingPointFormattingCSCodeAnalyzer())
+            var expected2 = DiagnosticResult.CompilerWarning(Descriptors.LuceneDev1002_FloatingPointArithmetic.Id)
+                .WithMessageFormat(Descriptors.LuceneDev1002_FloatingPointArithmetic.MessageFormat)
+                .WithArguments("(double)float1 * (double)float2")
+                .WithLocation("/0/Test0.cs", line: 17, column: 31);
+
+            var test = new InjectableCSharpAnalyzerTest(() => new LuceneDev1002_FloatingPointArithmeticCSCodeAnalyzer())
             {
                 TestCode = testCode,
-                ExpectedDiagnostics = { expected }
+                ExpectedDiagnostics = { expected1, expected2 }
             };
 
             await test.RunAsync();
